Move StandaloneView methods back into View
